Change Python file structure
