--- ISSUE_TEMPLATE.md
@@ -1,18 +1,0 @@
-### Excpected Behavior
-
-*what did you expect to happen based on your tests and the documentation. be descriptive.*
-
-### Actual Behavior
-
-*what acually happened when you ran the code. be descriptive.*
-
-### Operating Conditions
-
-*what operating system (OS) are you using? what version of Python? which version of the code? etc...*
-
-### Screenshots
-
-*attach any screenshots you might have taken here*
-
----
-@willtheorangeguy
