--- PULL_REQUEST_TEMPLATE.md
@@ -1,12 +1,0 @@
-Fixes issue # .
-
-Changes proposed in this PR:
-
--
--
--
-
-Reason for PR: *replace this text*
-
----
-@willtheorangeguy
